2018/08/09 9:31

+++ app/src/main/java/com/example/han/referralproject/intelligent_diagnosis/MonthlyReport4Fragment.java
@@ -136,7 +136,7 @@ public class MonthlyReport4Fragment extends Fragment {
 
         if (pc_yan > 0) {
             imgYan.setImageResource(R.drawable.red_up);
-            pcYan.setText((int) pc_yan + "");
+            pcYan.setText(pc_yan + "");
         } else {
             imgYan.setVisibility(View.GONE);
             pcYan.setText("√");
@@ -145,7 +145,7 @@ public class MonthlyReport4Fragment extends Fragment {
 
         if (pc_yundong < 0) {
             imgYundong.setImageResource(R.drawable.red_down);
-            pcYundong.setText((int) pc_yundong + "");
+            pcYundong.setText(pc_yundong + "");
         } else {
             imgYundong.setVisibility(View.GONE);
             pcYundong.setText("√");
@@ -154,7 +154,7 @@ public class MonthlyReport4Fragment extends Fragment {
 
         if (pc_tizhong > 0) {
             imgTizhong.setImageResource(R.drawable.red_up);
-            pcTizhong.setText((int) pc_tizhong + "");
+            pcTizhong.setText(pc_tizhong + "");
         } else {
             imgTizhong.setVisibility(View.GONE);
             pcTizhong.setText("√");
@@ -163,7 +163,7 @@ public class MonthlyReport4Fragment extends Fragment {
 
         if (pc_yinjiu > 0) {
             imgYinjiu.setImageResource(R.drawable.red_up);
-            pcYinjiu.setText((int) pc_yinjiu + "");
+            pcYinjiu.setText(pc_yinjiu + "");
         } else {
             imgYinjiu.setVisibility(View.GONE);
             pcYinjiu.setText("√");
